Added battle_common.h where seen fit
